fix clippy (rust 1.95.0) warnings

+++ rust/src/problems/p1901.rs
@@ -22,7 +22,7 @@ impl Solution {
             }
         }
         let j = max_index(&mat[l as usize]) as i32;
-        vec![l, j as i32]
+        vec![l, j]
     }
 }
